Silence platform-dependent compiler whining: in general, when printf() whines about some platform-dependent integer size issue, it's best to use both an explicitly sized format (eg, "%lu") and an explicit cast (eg, "(unsigned long)") when silencing the warning, otherwise it'll just pop up again in different form on the next platform tested.

+++ tests/test-bus.c
@@ -91,7 +91,8 @@ static void _time_check(char *label, const struct timeval t0, const int err)
     t.tv_sec  -= 1;
   }
   rounds = (float)TEST_NUM_ROUNDS/((float)t.tv_sec + ((float)t.tv_usec / 1000000));
-  printf("%s%lu.%06lu seconds, %u/sec\n", label, t.tv_sec, t.tv_usec, (unsigned)rounds);
+  printf("%s%lu.%06lu seconds, %u/sec\n", label,
+	 (unsigned long)t.tv_sec, (unsigned long)t.tv_usec, (unsigned)rounds);
 }
 
 #define time_check(_label_, _expr_)             \
